feature/testing: Splitting out unit and integration tests.
- Moving stuff out of main.rs into a library so things can be used in different crates.
