+++ src/main.rs
@@ -1,13 +1,11 @@
 #![no_std] // Disable stdlib linking
 #![no_main] // Disable all rust-level entry points
 #![feature(custom_test_frameworks)]
-#![test_runner(crate::test_runner)]
+#![test_runner(fractal_os::test_runner)]
 #![reexport_test_harness_main = "test_main"]
 
 use core::panic::PanicInfo;
-
-mod vga_buffer;
-mod serial;
+use fractal_os::{println, print};
 
 // Don't mangle the entry point function name
 #[no_mangle]
@@ -31,64 +29,10 @@ fn panic(info: &PanicInfo) -> ! {
     loop {}
 }
 
-// -------- Testing Functions --------
-
-// Test panic handler
 #[cfg(test)]
 #[panic_handler]
 fn panic(info: &PanicInfo) -> ! {
-    serial_println!("[failure]");
-    serial_println!();
-    serial_println!("*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-");
-    serial_println!("Tests Failed.");
-    serial_println!();
-    serial_println!("[KERNEL_PANIC]: {}", info);
-    serial_println!("*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-*-");
-    exit_qemu(QemuExitCode::Failed);
-    loop {}
-}
-
-#[derive(Debug, Copy, Clone, Eq, PartialEq)]
-#[repr(u32)]
-pub enum QemuExitCode {
-    Success = 0x10,
-    Failed = 0x11
-}
-
-pub fn exit_qemu(exit_code: QemuExitCode) {
-    use x86_64::instructions::port::Port;
-
-    unsafe {
-        let mut port = Port::new(0xf4);
-        port.write(exit_code as u32);
-    }
-}
-
-pub trait Testable {
-    fn run(&self) -> ();
-}
-
-impl<T> Testable for T where T: Fn() {
-    fn run(&self) {
-        serial_print!("{}...\t", core::any::type_name::<T>());
-        self();
-        serial_println!("[ok]");
-    }
-}
-
-#[cfg(test)]
-fn test_runner(tests: &[&dyn Testable]) {
-    serial_println!();
-    serial_println!("================================================");
-    serial_println!("Running {} tests", tests.len());
-    serial_println!("================================================");
-    serial_println!();
-
-    for test in tests {
-        test.run();
-    }
-
-    exit_qemu(QemuExitCode::Success);
+    fractal_os::test_panic_handler(info);
 }
 
 #[test_case]
