+++ src/vga_buffer.rs
@@ -157,3 +157,28 @@ impl fmt::Write for Writer {
         Ok(())
     }
 }
+
+// ----- Testing -----
+
+#[test_case]
+fn test_println_single_line() {
+    println!("test_println_single_line");
+}
+
+#[test_case]
+fn test_println_multiple_lines() {
+    for _ in 0..350 {
+        println!("test_println_multiple_lines");
+    }
+}
+
+#[test_case]
+fn test_println_output() {
+    let string = "test_println_output";
+    println!("{}", string);
+
+    for (i, c) in string.chars().enumerate() {
+        let screen_char = WRITER.lock().buffer.chars[BUFFER_HEIGHT - 2][i].read();
+        assert_eq!(char::from(screen_char.ascii_character), c);
+    }
+}
